move tjson under tigris handler
