+++ internal/handlers/tigris/json_schema.go
@@ -17,8 +17,9 @@ package tigris
 import (
 	"fmt"
 
+	"github.com/FerretDB/FerretDB/internal/handlers/tigris/tjson"
+
 	"github.com/FerretDB/FerretDB/internal/handlers/common"
-	"github.com/FerretDB/FerretDB/internal/tjson"
 	"github.com/FerretDB/FerretDB/internal/types"
 )
