+++ internal/handlers/tigris/msg_delete.go
@@ -19,11 +19,12 @@ import (
 	"encoding/json"
 	"fmt"
 
+	"github.com/FerretDB/FerretDB/internal/handlers/tigris/tjson"
+
 	"github.com/tigrisdata/tigris-client-go/driver"
 
 	"github.com/FerretDB/FerretDB/internal/handlers/common"
 	"github.com/FerretDB/FerretDB/internal/handlers/tigris/tigrisdb"
-	"github.com/FerretDB/FerretDB/internal/tjson"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/util/must"
