+++ internal/handlers/tigris/tigrisdb/insert.go
@@ -17,9 +17,10 @@ package tigrisdb
 import (
 	"context"
 
+	"github.com/FerretDB/FerretDB/internal/handlers/tigris/tjson"
+
 	"github.com/tigrisdata/tigris-client-go/driver"
 
-	"github.com/FerretDB/FerretDB/internal/tjson"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/util/must"
