+++ internal/handlers/tigris/tigrisdb/query.go
@@ -18,10 +18,11 @@ import (
 	"context"
 	"encoding/json"
 
+	"github.com/FerretDB/FerretDB/internal/handlers/tigris/tjson"
+
 	"github.com/tigrisdata/tigris-client-go/driver"
 	"go.uber.org/zap"
 
-	"github.com/FerretDB/FerretDB/internal/tjson"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/util/must"
