+++ internal/handlers/tigris/tjson/binary.go
